refactor: centralize gRPC contracts via Buf and shared library
