+++ go.work
@@ -1,6 +1,7 @@
 go 1.25.6
 
 use (
+	./services/gen/go
 	./services/weather-collector
 	./services/weather-provider
 )
