+++ services/gen/go/go.mod
@@ -1,0 +1,15 @@
+module github.com/nickfang/personal-dashboard/services/gen/go
+
+go 1.25.6
+
+require (
+	google.golang.org/grpc v1.79.0
+	google.golang.org/protobuf v1.36.11
+)
+
+require (
+	golang.org/x/net v0.48.0 // indirect
+	golang.org/x/sys v0.39.0 // indirect
+	golang.org/x/text v0.32.0 // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20251202230838-ff82c1b0f217 // indirect
+)
