+++ services/protos/buf.yaml
@@ -1,0 +1,9 @@
+# For details on buf.yaml configuration, visit https://buf.build/docs/configuration/v2/buf-yaml
+version: v2
+name: buf.build/nickfang/personal-dashboard
+lint: {
+  use: [STANDARD]
+}
+breaking: {
+  use: [FILE]
+}
